requirements.txt
aelog==1.0.3
aiofiles==0.4.0
aiohttp==3.5.4
aiomysql==0.0.20
aioredis==1.2.0
aredis==1.1.5
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
bcrypt==3.1.7
bson==0.5.8
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorlog==4.0.2
cryptography==2.7
h11==0.8.1
h2==3.1.1
hiredis==1.0.0
honcho==1.0.1
hpack==3.0.0
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
idna-ssl==1.1.0
motor==2.0.0
multidict==4.5.2
pycparser==2.19
PyJWT==1.7.1
pymongo==3.9.0
PyMySQL==0.9.2
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
requests-async==0.5.0
rfc3986==1.3.2
sanic==19.6.2
Sanic-Cors==0.9.8.post3
sanic-motor==0.3.4
Sanic-Plugins-Framework==0.8.2
sanic-session==0.5.6
sanicdb==0.2
six==1.12.0
SQLAlchemy==1.3.8
typing-extensions==3.7.4
ujson==1.35
urllib3==1.25.3
uvloop==0.12.2
websockets==6.0
yarl==1.3.0
